+++ src/Components/Items/Icons/NatronIcon.tsx
@@ -1,5 +1,5 @@
 import { SvgIcon, SvgIconProps } from "@mui/material";
-import { ReactComponent as NatronLogo } from "../../Assets/logo_big_color.svg";
+import { ReactComponent as NatronLogo } from "../../../Assets/logo_big_color.svg";
 
 export default function NatronIcon(props: SvgIconProps) {
   return <SvgIcon component={NatronLogo} inheritViewBox fontSize="inherit" />;
