+++ src/Components/Items/Icons/NatronIconWhite.tsx
@@ -1,5 +1,5 @@
 import { SvgIcon, SvgIconProps } from "@mui/material";
-import { ReactComponent as NatronLogoWhite } from "../../Assets/logo_big_white.svg";
+import { ReactComponent as NatronLogoWhite } from "../../../Assets/logo_big_white.svg";
 
 export default function NatronIconWhite(props: SvgIconProps) {
   return (
