+++ src/Components/Navbar.tsx
@@ -19,8 +19,8 @@ import SettingsIcon from "@mui/icons-material/Settings";
 import MonetizationOnIcon from "@mui/icons-material/MonetizationOn";
 import BarChartIcon from "@mui/icons-material/BarChart";
 import { Link } from "react-router-dom";
-import NatronIcon from "./Items/NatronIcon";
-import NatronIconWhite from "./Items/NatronIconWhite";
+import NatronIcon from "./Items/Icons/NatronIcon";
+import NatronIconWhite from "./Items/Icons/NatronIconWhite";
 import drawerContext from "./Items/DrawerContext";
 import { useContext } from "react";
 import TenantDropdown from "./Items/TenantDropdown";
